Fix: Display app icons alongside work profile/private space indicators
This commit modifies the `AppDrawerAdapter` to ensure that application icons are displayed in the app drawer even when work profile or private space indicators are also present.

Previously, if an app was part of a work profile or private space, its own icon would be replaced by the respective indicator icon. Now, both the app icon and the indicator icon are shown.

The logic for setting compound drawables has been updated to accommodate this change. If app icons are enabled, the code now checks if a work profile or private space indicator needs to be shown and positions them appropriately next to the app icon.

Signed-off-by: CreativeCodeCat <wayne6324@gmail.com>