feat: Add support for alpha/beta version suffixes (Google Play)
This commit introduces the ability to append "-alpha" or "-beta"
suffixes to the version name based on a new `type` variable in
`app/build.gradle.kts`.

If `type` is set to 1, "-beta" is appended.
If `type` is set to 2, "-alpha" is appended.
Otherwise, no suffix is added.

The `baseVersionName` is constructed from `major`, `minor`, `patch`,
and `build` values, and the `versionNameStr` is then derived from
this base and the `type`.

Signed-off-by: CreativeCodeCat <wayne6324@gmail.com>