Feat(UI): Add action FAB to HomeFragment
This commit introduces a new Floating Action Button (FAB) named `fabAction` to the `HomeFragment`.

The `fabAction` is now included in the `fabList`, which is used to manage the visibility and color of the FABs based on user preferences. This allows users to customize the action FAB alongside other existing FABs (phone, messages, camera, photos, browser, settings).

Signed-off-by: CreativeCodeCat <wayne6324@gmail.com>