fix: correct DroidSans font mapping
The DroidSans font was incorrectly mapped to OpenSans. This commit corrects the mapping to use the actual DroidSans font file.

It also appears that all font files in `app/src/main/res/font/` were updated, likely due to a reformatting or re-saving process, though the content of these files has not substantially changed.

Signed-off-by: CreativeCodeCat <wayne6324@gmail.com>