Refactor: Log activities from package in MainViewModel
This commit introduces a call to `logActivitiesFromPackage` within the `launchApp` function of `MainViewModel.kt`. This will log all activities associated with the package name of the app being launched.

The purpose of this change is to aid in debugging and understanding app behavior by providing a record of launched activities.

Signed-off-by: CreativeCodeCat <wayne6324@gmail.com>