Refactor: Use FragmentActivity for BiometricHelper
The BiometricHelper now takes a FragmentActivity as a constructor parameter instead of a Fragment.

This change simplifies the initialization of BiometricHelper in various parts of the app:

- AppDrawerAdapter: Now passes `fragment.requireActivity()`
- HomeFragment: Now passes `this.requireActivity()`
- MainViewModel: Now passes `fragment.requireActivity()`

Additionally, in MainViewModel, the `hideKeyboard()` call for locked apps is now deferred until after successful biometric authentication and only executed if the fragment is still added.

Signed-off-by: CreativeCodeCat <wayne6324@gmail.com>