feat: Disable inclusion of dependency info in APKs and Bundles
This commit updates the `app/build.gradle.kts` file to explicitly disable the inclusion of dependency information (VCS information for software dependencies) in both APKs and Bundles.

The `dependenciesInfo` block has been added with `includeInApk` and `includeInBundle` set to `false`. This can help reduce the size of the generated artifacts and prevent the accidental inclusion of potentially sensitive version control system information.

Signed-off-by: CreativeCodeCat <wayne6324@gmail.com>